+++ server/src/entity/entity.rs
@@ -1,0 +1,11 @@
+use crate::util::math::DbVector3;
+
+#[spacetimedb::table(name = entity, public)]
+#[derive(Debug, Clone)]
+pub struct Entity {
+    #[auto_inc]
+    #[primary_key]
+    pub entity_id: u32,
+    pub position: DbVector3,
+    pub mass: u32,
+}
