+++ server/src/entity/mod.rs
@@ -1,0 +1,3 @@
+pub mod cow;
+pub mod ufo;
+pub mod entity;
