+++ server/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod util;
-pub mod entities;
-pub mod reducers;
-pub mod math;
-pub mod constants;
+pub mod system;
+pub mod player;
+pub mod entity;
