+++ server/src/util/math.rs
